+++ src/lib/jp2/util/util.h
@@ -87,7 +87,7 @@ struct grk_buf {
 	uint8_t* curr_ptr();
 
 	uint8_t *buf;		/* internal array*/
-    uint64_t offset;	/* current offset into array */
+    size_t offset;	/* current offset into array */
     size_t len;		/* length of array */
     bool owns_data;	/* true if buffer manages the buf array */
 } ;
